+++ src/test/java/de/learnlib/ralib/example/list/BoundedListDataWordOracle.java
@@ -1,5 +1,6 @@
 package de.learnlib.ralib.example.list;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.function.Supplier;
 
@@ -53,13 +54,13 @@ public class BoundedListDataWordOracle implements DataWordOracle {
 	private boolean getOutput(PSymbolInstance symInst, BoundedList list) {
 		try {
 			if (symInst.getBaseSymbol().equals(PUSH)) {
-				list.push((Integer) symInst.getParameterValues()[0].getId());
+				list.push((BigDecimal) symInst.getParameterValues()[0].getId());
 				return true;
 			} else if (symInst.getBaseSymbol().equals(POP)) {
-				Integer value = list.pop();
+				BigDecimal value = list.pop();
 				return symInst.getParameterValues()[0].getId().equals(value); 
 			} else if (symInst.getBaseSymbol().equals(INSERT)) {
-				list.insert((Integer) symInst.getParameterValues()[0].getId(), (Integer) symInst.getParameterValues()[1].getId());
+				list.insert((BigDecimal) symInst.getParameterValues()[0].getId(), (BigDecimal) symInst.getParameterValues()[1].getId());
 				return true;
 			} else if (symInst.getBaseSymbol().equals(CONTAINS)) {
 				return list.contains((Integer) symInst.getParameterValues()[0].getId());
